Remnant template folder from trying to clean up
Created a branch called simple-form-gem and did a bunch of work, and
then went back to master to create another branch, assuming master would
be unscathed from the simple-form-gem branch changes.  It was not.  I
did not add or commit the changes from the simple-form-gem branch, but
they were there on master, and I did not know how to create a new third
branch that did not contain these changes.  I figured something out, but
decided to leave this template file because I was unsure what happened
with it.