Handmade coding option for error presentation
Tried my hand at the handmade error option from the Validating Posts
assignment.  Cannot get it to work, completely.  I modified the method
in the application_helper file so that it is basically just a
conditional statement returning what the div's class values will be.  I
then updated the _form.html.erb file with the new method's name, and
tried to update the div's class to accept what the method will be
returning, but it does not work 100%.  It returns the first class value
of 'form-group'.  It does not include the second class value of
'has-error' properly.  It does return it, but I can't seem to get the
'has-error' to be in the class name.  Right now when I inspect the
element in the browser, it looks like this:  <div class"form-group"
has-error>.  So the 'has-error' is outside of the quotes.  Maybe it has
something to do with safe html, or raw, or something, but I can't figure
it out.