Comments does not work
Committing this branch now so I can create a clean branch off of master
that does not contain any of my changes that were made in the Comments
branch. My understanding is that I need to push this branch to GitHub
for this to work.  Otherwise, if I just checkout master without
committing and pushing the comments branch, master will contain this
changes.