Completed the comments checkpoint
Got the comments working.  Had difficulties understanding the logic for
posts_controller/show action and the comments_controller/create action.
Also line 4 of the comments/_form.html.erb file is not crystal clear.
Added the validations to the comment model, but did not created a
dedicated way for those errors to display so the user knows what they
did wrong.