In posts/_form I added "" around class erb tag
And I removed the sets_div_class_value method from the body's
f.text_area line.  It was only in there for practice, when I was trying
to figure out a solution to get this all to work.  Nonetheless, I
thought I tried putting thie "" around the erb tags in the <div class>
for title, but apparently I didn't.  Good catch.  Thank you.