Fixed the save_with_initial_vote method
By not passing in a parameter.  Instead, calling the method on the @post
object.  Also by recognizing the self is implicit in the
save_with_initial_vote method; therefore, this method can just call save
and create_vote, and by adding those two methods, it is doing it on
itself, because calling them in this fashion is implying thy are being
called on self.